Restructure authentication for LDAP
This patch removes the dependancy on the obscure pam client module in
favor of the mainline nis and crypt modules. Additionally, LDAP is
presented as the primary authentication technique. NIS is only used if
LDAP failes. I've added the option to migrate passwords to LDAP if
LDAP fails while NIS succeeds.